refactor: moved stac catalog explorer to components folder
