+++ sites/geohub/src/components/util/stac/StacCatalogItem.svelte
