+++ sites/geohub/src/components/util/stac/StacCollectionMap.svelte
