+++ sites/geohub/src/routes/(app)/management/+page.ts
@@ -1,0 +1,11 @@
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async () => {
+	const title = 'Management | GeoHub';
+	const content = 'GeoHub Management';
+
+	return {
+		title,
+		content
+	};
+};
