+++ sites/geohub/src/routes/(app)/management/stac/+page.ts
@@ -1,0 +1,11 @@
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async () => {
+	const title = 'STAC management | GeoHub';
+	const content = 'STAC management';
+
+	return {
+		title,
+		content
+	};
+};
